Add files via upload
This is a Java based project which is an Airline Booking System. Users can login and book a flight only across the locations offered within several classes simplifying the booking process and logistics for a particular region. And ticket can be downloaded as the text file. The project involves the use of various concepts such as GUI (Swing and AWT), exception handling, file handling, JDBC, etc.